Use ITRS and astropy's EarthLocation mechanism for array center coordinates
This description of the coordinate system in the OIFITS 1 and 2 standard
papers is vague; adopt the same standard (i.e. ITRS) as uvfits. This is
an incompatible, non-breaking change with how the module calculated
array height before, so the array height values reported from ofits
files written with previous versions of this module will be incorrect.

The module tries to check if the files were written with an earlier
version and computes a new arrxyz if necessary.